An Online 2-User X/O Game
A simple 2-user game to play with your peers whenever you are bored! Made using only HTML, CSS and Javascript, this site not only tells us whose turn it is but also tells us the winner and comes along with the feature of Reset to play as many times as you want to.
This project is Responsive as well hence, it can be played across mobiles and various other devices as well. Additional features include Music played at each turn and on the game getting over along with a line drawn on the winner’s pattern with an added animation.